BroadleafCommerce/QA#4262 Move sorting out from the getter
